grpc_communication/client/requirements.txt
grpcio==1.56.0
grpcio-tools==1.56.0
protobuf==4.25.2
